api/requirements.txt
flask
requests
python-dotenv
flask-cors
pymongo
